+++ js/util.js
@@ -1,0 +1,4 @@
+const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1));
+
+
+export { getRandomNumber };
